#2144 - Project state change events only work for matrix workload
- Pull out code for updating the document state to the dao module because it is the same for all (existing) workflow types
- Pull code for obtaining source document stats out into the document service
- Undeprecate the ANNOTATION_FINISHED source document state
