+++ inception/inception-api/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/DocumentService.java
@@ -690,4 +690,6 @@ public interface DocumentService
     void upgradeAllAnnotationDocuments(Project aProject) throws IOException;
 
     Map<AnnotationDocumentState, Long> getAnnotationDocumentStats(SourceDocument aDocument);
+
+    SourceDocumentStateStats getSourceDocumentStats(Project aProject);
 }
