+++ inception/inception-api/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/SourceDocumentStateStats.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.project;
+package de.tudarmstadt.ukp.clarin.webanno.api;
 
 import de.tudarmstadt.ukp.clarin.webanno.model.ProjectState;
