+++ inception/inception-workload-matrix/src/main/java/de/tudarmstadt/ukp/inception/workload/matrix/event/MatrixWorkloadUpdateDocumentStateTask.java
@@ -1,3 +1,20 @@
+/*
+ * Licensed to the Technische Universität Darmstadt under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The Technische Universität Darmstadt 
+ * licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.
+ *  
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package de.tudarmstadt.ukp.inception.workload.matrix.event;
 
 import static de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState.ANNOTATION_FINISHED;
@@ -23,7 +40,7 @@ import de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState;
 import de.tudarmstadt.ukp.inception.scheduling.DebouncingTask;
 import de.tudarmstadt.ukp.inception.workload.model.WorkloadManagementService;
 
-public class MatrixUpdateDocumentStateTask
+public class MatrixWorkloadUpdateDocumentStateTask
     extends DebouncingTask
 {
     private @PersistenceContext EntityManager entityManager;
@@ -33,7 +50,7 @@ public class MatrixUpdateDocumentStateTask
 
     private final SourceDocument document;
 
-    public MatrixUpdateDocumentStateTask(SourceDocument aDocument, String aTrigger)
+    public MatrixWorkloadUpdateDocumentStateTask(SourceDocument aDocument, String aTrigger)
     {
         super(aDocument.getProject(), aTrigger, 15_000l);
         document = aDocument;
@@ -53,6 +70,8 @@ public class MatrixUpdateDocumentStateTask
             return;
         }
 
+        // We check this here instead of checking at task submission to avoid hammering the
+        // DB if there is a high event frequency
         if (!MATRIX_WORKLOAD_MANAGER_EXTENSION_ID.equals(workloadManagementService
                 .loadOrCreateWorkloadManagerConfiguration(project).getType())) {
         }
@@ -94,7 +113,7 @@ public class MatrixUpdateDocumentStateTask
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        MatrixUpdateDocumentStateTask task = (MatrixUpdateDocumentStateTask) o;
+        MatrixWorkloadUpdateDocumentStateTask task = (MatrixWorkloadUpdateDocumentStateTask) o;
         return document.equals(task.document) && getProject().equals(task.getProject());
     }
